some more decomp

+++ include/code_80130A8.h
@@ -11,7 +11,7 @@ extern void sub_8013F84(void);
 extern void sub_80140B4(UnkTextStruct2 *);
 extern void sub_80140DC(void);
 extern void sub_8014114(void);
-extern void sub_80141B4(const u8 *, u32, OpenedFile **, u32);
+extern void sub_80141B4(const u8 *, u32, OpenedFile **, u16);
 extern void sub_8014248(const u8 *, u32, u32, const MenuItem *, void *, u32, u32, OpenedFile **, u32);
 extern s32 sub_80144A4(s32 *);
 extern void sub_80151C0(u32, u8 *);
